+++ example/app/bootstrap/bootstrap-example.component.ts
@@ -42,7 +42,7 @@ export class BootstrapExampleComponent implements OnInit {
 
         this.arrayControl = <FormArray> this.formGroup.get("bootstrapFormGroup2").get("bootstrapFormArray");
         this.arrayModel = <DynamicFormArrayModel> this.formService.findById(
-            "bootstrapFormArray", this.formModel);
+            "bootstrapFormArray", (<DynamicFormGroupModel> this.formModel[1]).group);
     }
 
     add() {
@@ -57,14 +57,20 @@ export class BootstrapExampleComponent implements OnInit {
         this.formService.removeFormArrayGroup(index, this.arrayControl, context);
     }
 
+    move (context: DynamicFormArrayModel, index: number, step: number) {
+        this.formService.moveFormArrayGroup(index, step, this.arrayControl, context);
+    }
+
     clear() {
         this.formService.clearFormArray(this.arrayControl, this.arrayModel);
     }
 
     test() {
         //this.exampleModel.disabledUpdates.next(!this.exampleModel.disabled);
-        this.exampleModel.valueUpdates.next("Hello Hello");
+        //this.exampleModel.valueUpdates.next("Hello Hello");
         //console.log(JSON.stringify(this.exampleModel));
+        //this.arrayModel.get(1).group[0].valueUpdates.next("This is just a test");
+        this.formService.moveFormArrayGroup(2, -1, this.arrayControl, this.arrayModel);
     }
 
     onBlur($event) {