+++ modules/core/src/service/dynamic-form.service.spec.ts
@@ -208,7 +208,7 @@ describe("DynamicFormService test suite", () => {
 
     it("should create a form array correctly", () => {
 
-        let model = service.findById("testFormArray", testModel),
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
             formArray;
 
         expect(service.createFormArray).toBeDefined();
@@ -220,18 +220,9 @@ describe("DynamicFormService test suite", () => {
     });
 
 
-    it("should create a form array group correctly", () => {
-
-        let model = service.findById("testFormArray", testModel);
-
-        expect(service.createFormArrayGroup).toBeDefined();
-        expect(service.createFormArrayGroup(model) instanceof FormGroup).toBe(true);
-    });
-
-
     it("should add a form array group correctly", () => {
 
-        let model = service.findById("testFormArray", testModel),
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
             formArray = service.createFormArray(model);
 
         expect(service.addFormArrayGroup).toBeDefined();
@@ -244,7 +235,7 @@ describe("DynamicFormService test suite", () => {
 
     it("should insert a form array group correctly", () => {
 
-        let model = service.findById("testFormArray", testModel),
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
             formArray = service.createFormArray(model);
 
         expect(service.insertFormArrayGroup).toBeDefined();
@@ -255,9 +246,47 @@ describe("DynamicFormService test suite", () => {
     });
 
 
+    it("should move a form array group correctly", () => {
+
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
+            formArray = service.createFormArray(model),
+            formArrayGroup,
+            index = 3,
+            step = -2;
+
+        expect(service.moveFormArrayGroup).toBeDefined();
+
+        formArray.at(index).controls["basicArrayGroupInput"].setValue("next test value 1");
+        formArray.at(index + step).controls["basicArrayGroupInput"].setValue("next test value 2");
+
+        model.get(index).get(0).valueUpdates.next("next test value 1");
+        model.get(index + step).get(0).valueUpdates.next("next test value 2");
+
+        service.moveFormArrayGroup(index, step, formArray, model);
+
+        expect(formArray.length).toBe(model.initialCount);
+
+        expect(formArray.at(index).controls["basicArrayGroupInput"].value).toEqual("next test value 2");
+        expect(formArray.at(index + step).controls["basicArrayGroupInput"].value).toEqual("next test value 1");
+
+        expect(model.get(index).get(0).value).toEqual("next test value 2");
+        expect(model.get(index + step).get(0).value).toEqual("next test value 1");
+    });
+
+
+    it("should throw when form array group is to be moved out of bounds", () => {
+
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
+            formArray = service.createFormArray(model);
+
+        expect(() => service.moveFormArrayGroup(2, -5, formArray, model))
+            .toThrow(new Error(`form array group cannot be moved due to index or new index being out of bounds`));
+    });
+
+
     it("should remove a form array group correctly", () => {
 
-        let model = service.findById("testFormArray", testModel),
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
             formArray = service.createFormArray(model);
 
         expect(service.removeFormArrayGroup).toBeDefined();
@@ -270,7 +299,7 @@ describe("DynamicFormService test suite", () => {
 
     it("should clear a form array correctly", () => {
 
-        let model = service.findById("testFormArray", testModel),
+        let model = <DynamicFormArrayModel> service.findById("testFormArray", testModel),
             formArray = service.createFormArray(model);
 
         expect(service.clearFormArray).toBeDefined();